+++ docs/changelog.rst
@@ -5,3 +5,9 @@ Première version de la librairie
 
 * Décorateur de classe permettant de serialiser des modèles SQLAlchemy via la méthode `as_dict` (recursivité, choix de colonnes, choix de relationships)
 * Fonctions utilitaires pour retourner des réponses HTTP JSON ou CSV.
+
+0.0.2generic (2019-22-17)
+-------------------------
+
+* ajout des Generic Table et GenericQuery (en version simplifiée sans la gestion des géométries)
+* ajout des exceptions GeonatureApiError